Updates migrations
Changes datatype for node lat and lon;
Changes name of bike atribue on tags;
Removes 4wd_only tag;

+++ db/migrations/20180525094634_create-nodes-table.js
@@ -2,8 +2,8 @@
 exports.up = function(knex, Promise) {
   return knex.schema.createTable('nodes', function(t) {
     t.increments();
-    t.integer('latitude');
-    t.integer('longitude');
+    t.float('latitude');
+    t.float('longitude');
     t.integer('trail_id').references('trails.id');
     t.timestamps();
   });
